fix: personal vips can only have savings account

+++ src/main/java/com/project1/ms_account_service/business/factory/CheckingAccountCreator.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 public class CheckingAccountCreator implements AccountCreator {
 
     @Value("${account.config.checking.maintenanceFee}")
-    private Double checkingAccountMaintenanceFee;
+    private BigDecimal checkingAccountMaintenanceFee;
 
     @Value("${account.config.checking.maxMonthlyMovementsNoFee}")
     private Integer checkingMaxMonthlyMovementsNoFee;
